requirements.txt
openai
chromadb
python-dotenv
beautifulsoup4
requests
fastapi
uvicorn
streamlit
crewai

networkx
matplotlib
firecrawl-py